Made changes to AccountController.cs and signin.html
Not sure what is going on but for some reason the javascript at the
bottom is not executing at all. I currently have the AccountController's
login function taking in the username, password, and boolean remember
me. However as of right now I am not sure how to hook up the controller
to the UserDao file, so it is calling the WebSecurity's Login method.